updated the bash scripts for deployment and delete

+++ delete.sh
@@ -1,11 +1,21 @@
 #!/bin/bash
 
-sudo rm /etc/nginx/sites-enabled/kyodo-backend
+echo "This script will undo the changes made by the previous script. Are you sure you want to continue? [Y/n]"
+read response
+upperstr=$(echo $response | tr '[:lower:]' '[:upper:]')
 
-sudo systemctl restart nginx
-sudo systemctl status nginx
+if [[ "$response" = "" || "$upperstr" = "Y" ]]
+then
+    sudo rm /etc/nginx/sites-enabled/kyodo-backend
+    sudo rm /etc/nginx/sites-available/kyodo-backend
 
-pm2 delete kyodo-backend
-pm2 save --force
+    sudo systemctl restart nginx
+    sudo systemctl status nginx
 
-echo "Everything deleted"
+    pm2 delete kyodo-backend
+    pm2 save --force
+
+    echo "Everything deleted"
+else
+    echo "Exiting without making any changes."
+fi
